Rename component to reflect year
