+++ src/components/Organisms/Header2025/Header2025.js
@@ -2,31 +2,31 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import Logos from '../../Molecules/Logos/Logos';
-import HeaderNav from './HeaderNav2024/HeaderNav2024';
+import HeaderNav from './HeaderNav2025/HeaderNav2025';
 import {
-  Brand, Header2024Wrapper, InnerWrapper,
-  DonateButtonWrapperTop, Header2024MetaIcons, ButtonsAndIcons
-} from './Header2024.style';
+  Brand, Header2025Wrapper, InnerWrapper,
+  DonateButtonWrapperTop, Header2025MetaIcons, ButtonsAndIcons
+} from './Header2025.style';
 
-const Header2024 = ({
+const Header2025 = ({
   navItems = {}, metaIcons, campaign = 'Comic Relief', donateButton = null, ...rest
 }) => (
-  <Header2024Wrapper navItems {...rest}>
+  <Header2025Wrapper navItems {...rest}>
     <InnerWrapper>
       <Brand>
         <Logos campaign={campaign} sizeSm="100px" sizeMd="100px" />
       </Brand>
       <HeaderNav navItems={navItems} metaIcons={metaIcons} donateButton={donateButton} />
       <ButtonsAndIcons>
-        <Header2024MetaIcons isHeader data-testid="meta-icons--desktop">{metaIcons}</Header2024MetaIcons>
+        <Header2025MetaIcons isHeader data-testid="meta-icons--desktop">{metaIcons}</Header2025MetaIcons>
         <DonateButtonWrapperTop data-testid="donate-button--desktop">{donateButton}</DonateButtonWrapperTop>
       </ButtonsAndIcons>
 
     </InnerWrapper>
-  </Header2024Wrapper>
+  </Header2025Wrapper>
 );
 
-Header2024.propTypes = {
+Header2025.propTypes = {
   // Check data structure example in src/components/moleculecules/header/data/data
   navItems: PropTypes.objectOf(PropTypes.shape),
   // NB: metaIcons no longer include the Donate button:
@@ -36,4 +36,4 @@ Header2024.propTypes = {
   campaign: PropTypes.string
 };
 
-export default Header2024;
+export default Header2025;
