+++ src/components/Organisms/Header2025/HeaderNav2025/MoreNav.style.js
@@ -1,7 +1,7 @@
 import styled, { css } from 'styled-components';
 import {
   NavItem, SubNavMenu, NavLink, SubNavItem, SubNavLink
-} from './HeaderNav2024.style';
+} from './HeaderNav2025.style';
 
 const MoreSubNavMenu = styled(SubNavMenu)`
   @media ${({ theme }) => theme.allBreakpoints('Nav')} {
