+++ src/index.js
@@ -42,7 +42,7 @@ export { default as CardDs } from './components/Molecules/CardDs/CardDs';
 export { default as Box } from './components/Molecules/Box/Box';
 export { default as ArticleTeaser } from './components/Molecules/ArticleTeaser/ArticleTeaser';
 export { default as Header } from './components/Organisms/Header/Header';
-export { default as Header2024 } from './components/Organisms/Header2024/Header2024';
+export { default as Header2025 } from './components/Organisms/Header2025/Header2025';
 export { default as Donate } from './components/Organisms/Donate/Donate';
 export { default as DoubleCopy } from './components/Molecules/DoubleCopy/DoubleCopy';
 export { default as PartnerLink } from './components/Molecules/PartnerLink/PartnerLink';
